sched/fair: Collect load-balancing stats for EQOS
Determine whether at the back of runqueues of a scheduling group is a task
of type EQOS_MAX_EFFICIENCY. Later in load balance this type of tasks will
be considered to be moved to CPUs with better energy efficiency.

Discussion:
Under this approach we look for EQOS setting of tasks only during load
balance. Also, we only look at the back of the queue. We miss tasks ahead
in the queue. This should not be a problem because if there is more than
one runnable task on the queue the group will be overloaded and we ignore
EQOS considerations.

A different approach is to have EE tasks tag themselves and their runqueue
as candidates for load balance. The load balancer would look for those
tags. This is similar to the handling of misfit tasks in topologies with
asymmetric capacity.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>