sched/fair: Skip ASYM_PACKING balancing for queues with EE EQOS task
Do not trigger idle load balance if we find that the current CPU is busy
and has an EE EQOS task at the back of the queue. We know that the task is
in the optimal place.

TODO: What to do if there are non-EE EQOS tasks in the queue? It would not
matter if all CPUs are busy. It may be, however, that the CPU is overloaded
while other CPUs are idle. It will not matter because we unconditionally
trigger idle load balance if nr_running > 1.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>