sched/fair: Let idle siblings of SMT cores help overloaded sched groups
The siblings of an SMT core share hardware resources. When the busiest
group has the prefer_sibling flag set, the imbalance calculation considers
this fact and the result is done using cores, not logical CPUs. However,
this is insufficient to relieve a busiest group of type group_overloaded
(one that has more running tasks than cores).

The calculated imbalance would be zero if the busiest group has a slightly
more running tasks than cores (For a busiest group of 4 cores and a local
group of 1 core, between 5 and 7 running tasks).

The idle siblings still have resources that can be put to work even if
other siblings are busy. Thee idle siblings can help the overloaded busiest
group.

Let the SMT core help the overloaded group by pulling as many tasks as
necessary to relieve the overloaded condition or the number of idle
siblings in the SMT core is exhausted.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>