sched/fair: Check EE tasks when considering on asymmetric active balance
The best placement for EE tasks is on lower-priority CPUs. Let a lower-
priority proceed with active balance.

If the source CPU is in an SMT core with more than one sibling busy we
still want do load balance. The destination CPU is idle. No need to
check for this condition.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>